Abstracted out long lines
